requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.2
anyio==3.7.1
attrs==25.3.0
black==23.12.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
fastapi==0.104.1
filelock==3.18.0
flake8==7.2.0
frozenlist==1.6.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
loguru==0.7.2
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.4.4
mypy==1.16.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
pytest-asyncio==1.0.0
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
anthropic==0.18.1
